Refreshed Horizon theme
Our Horizon branding refreshed based on the "Paper" Bootswatch theme.
Also changes the way the path to the theme was handled.

Mind that this only includes variables/styles that were overwritten from
the base theme.

I left the old `params.pp` even though we don't use it - let me know if
you want me to remove it or move system-specific part of the path there.

Please comment if you have suggestions regarding the feel and look of
the theme and you want me to change anything.

Part of PD-2992.